refactor(inflection): move fixture builders out of runtime
