+++ website/docs/contributing/locale-yaml-reference.mdx
@@ -179,7 +179,7 @@ src/Humanizer.SourceGenerators/Common/LocaleYamlCatalog.cs fa14af4899e697caf4353
 src/Humanizer.SourceGenerators/Common/ProfileDefinitions.cs 49d6775d1664050ec814e87dfe7051479be5d89417e236fd3b932d0d83b01463
 src/Humanizer.SourceGenerators/Generators/LocaleRegistryInput.cs a26610c0bcf095e73925b0b1a3878f1f403876de78be029981764c8e984dda5b
 src/Humanizer.SourceGenerators/Generators/ProfileCatalogs/FormatterProfileCatalogInput.cs a5e1bc7acbf5e831e228295f05a65ff8e640a88911ae1ab8a48e688362b2e977
-src/Humanizer.SourceGenerators/Generators/ProfileCatalogs/InflectionCatalogInput.cs c1e2f4a139fb01a938650256ef6933aac4918e968dd3805b3432b93b4786ff12
+src/Humanizer.SourceGenerators/Generators/ProfileCatalogs/InflectionCatalogInput.cs 6486bfff5fa259c95e31c1b6c402cefc33ef4f83a1f33e4a8e85c147ff533d93
 src/Humanizer.SourceGenerators/Generators/ProfileCatalogs/LocaleDurationCaseTableCatalogInput.cs 3c59b5cd89fe57deaef1f5aa82e5268302c68734c8df79887f124d818b3992f8
 src/Humanizer.SourceGenerators/Generators/ProfileCatalogs/LocalePhraseTableCatalogInput.cs be2d2728de4b54068ac0d88c32223b692ab7735233c15238fe129a30bebf2599
 src/Humanizer.SourceGenerators/Generators/ProfileCatalogs/NumberToWordsEngineContractFactory.cs d62ea9d94b5c74104a26cfddc9a1269481410867b44d39f707e10e25b1a556ff
